refactor(app): move stripe client to pkg
