+++ openmeter/appstripe/service.go
@@ -3,6 +3,7 @@ package appstripe
 import (
 	"context"
 
+	stripeclient "github.com/openmeterio/openmeter/openmeter/appstripe/client"
 	appstripeentity "github.com/openmeterio/openmeter/openmeter/appstripe/entity"
 )
 
@@ -11,5 +12,5 @@ type Service interface {
 }
 
 type AppService interface {
-	CreateCheckoutSession(ctx context.Context, input appstripeentity.CreateCheckoutSessionInput) (appstripeentity.StripeCheckoutSession, error)
+	CreateCheckoutSession(ctx context.Context, input appstripeentity.CreateCheckoutSessionInput) (stripeclient.StripeCheckoutSession, error)
 }
