+++ openmeter/appstripe/service/app.go
@@ -5,14 +5,15 @@ import (
 	"fmt"
 
 	"github.com/openmeterio/openmeter/openmeter/appstripe"
+	stripeclient "github.com/openmeterio/openmeter/openmeter/appstripe/client"
 	appstripeentity "github.com/openmeterio/openmeter/openmeter/appstripe/entity"
 )
 
 var _ appstripe.AppService = (*Service)(nil)
 
-func (s *Service) CreateCheckoutSession(ctx context.Context, input appstripeentity.CreateCheckoutSessionInput) (appstripeentity.StripeCheckoutSession, error) {
+func (s *Service) CreateCheckoutSession(ctx context.Context, input appstripeentity.CreateCheckoutSessionInput) (stripeclient.StripeCheckoutSession, error) {
 	if err := input.Validate(); err != nil {
-		return appstripeentity.StripeCheckoutSession{}, appstripe.ValidationError{
+		return stripeclient.StripeCheckoutSession{}, appstripe.ValidationError{
 			Err: fmt.Errorf("error create checkout session: %w", err),
 		}
 	}
